Organize multi vehicles (#2473)
* source files moved to multi dir

* python scripts moved to multi dir

* short README for the multi directory

* correct paths to multi
