+++ sw/airborne/modules/multi/README
@@ -1,0 +1,3 @@
+ctc: collective tracking control
+dcf: distributed circular formation control
+fc_rotor: formation control for rotorcraft
