+++ sw/ground_segment/python/multi/collective_tracking_control/ctcInitTables.py
