+++ sw/ground_segment/python/multi/fc_rotor/fc_rotor.py
