+++ sw/ground_segment/python/multi/fc_rotor/lib_rigid_formations.py
